Create three-column floating CSS layou
This will cause the three columns to float side by side, with each column taking up 32% of the width of the container. You can adjust the widths of the columns as needed to create the desired layout.

It's worth to note that using float property can lead to some layout issues especially with a responsive design, it's recommended to use Flexbox or Grid layout instead.